Moved Header, Footer and SocialButtons into folders
