+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import Header from "./components/Header.vue";
-import Footer from "./components/Footer.vue";
+import Header from "./components/Layout/Header.vue";
+import Footer from "./components/Layout/Footer.vue";
 import Content from "./components/Content/Content.vue";
 </script>
